+++ modules/graphdb/variables.tf
@@ -28,6 +28,11 @@ variable "resource_group_name" {
 
 # Networking
 
+variable "vmss_dns_servers" {
+  description = "List of DNS servers for the VMSS"
+  type        = list(string)
+}
+
 variable "virtual_network_id" {
   description = "Virtual network the DNS will be linked to"
   type        = string
@@ -75,8 +80,8 @@ variable "app_configuration_id" {
   type        = string
 }
 
-variable "app_configuration_name" {
-  description = "Name of the App Configuration store for GraphDB"
+variable "app_configuration_endpoint" {
+  description = "Endpoint of the App Configuration store for GraphDB"
   type        = string
 }
 
@@ -294,3 +299,10 @@ variable "nat_gateway_pip_idle_timeout" {
   type        = number
   default     = 5
 }
+
+# Customer provided user data scripts
+variable "user_supplied_scripts" {
+  description = "Array of additional shell scripts to execute sequentially after the templated user data shell scripts."
+  type        = list(string)
+}
+